requirements.txt
openai>=0.26.4
python-dotenv>=0.21.1
python-telegram-bot>=20.0
